demo-environment/requirements.txt
Flask>=2.0.0
PyYAML>=6.0